+++ sources/Application/WebPage/iTopWebPage.php
@@ -8,7 +8,7 @@
 use Combodo\iTop\Application\TwigBase\Twig\TwigHelper;
 use Combodo\iTop\Application\UI\Base\Component\Alert\AlertUIBlockFactory;
 use Combodo\iTop\Application\UI\Base\Component\Breadcrumbs\Breadcrumbs;
-use Combodo\iTop\Application\UI\Base\Component\Dialog\DialogDoNotShowAgainOptionBlock;
+use Combodo\iTop\Application\UI\Base\Component\Dialog\DoNotShowAgainOptionBlock;
 use Combodo\iTop\Application\UI\Base\Component\Panel\PanelUIBlockFactory;
 use Combodo\iTop\Application\UI\Base\Component\Template\TemplateUIBlockFactory;
 use Combodo\iTop\Application\UI\Base\iUIBlock;
@@ -898,7 +898,7 @@ HTML;
 		$aData['aLayouts']['oPageContent'] = $this->GetContentLayout();
 		$aData['aDeferredBlocks']['oPageContent'] = $this->GetDeferredBlocks($this->GetContentLayout());
 		// - Prepare generic templates
-		$aData['aTemplates']['oTemplateDoNotShowAgainOption'] = TemplateUIBlockFactory::MakeForBlock('ibo-dialog-option--do-not-show-again-template', new DialogDoNotShowAgainOptionBlock());
+		$aData['aTemplates']['oTemplateDoNotShowAgainOption'] = TemplateUIBlockFactory::MakeForBlock('ibo-dialog-option--do-not-show-again-template', new DoNotShowAgainOptionBlock());
 
 		// - Retrieve layouts linked files
 		//   Note: Adding them now instead of in the template allow us to remove duplicates and lower the browser parsing time
